+++ chronometer.h
@@ -4,9 +4,9 @@
 
 class Chronometer {
     private:
-        int second;
-        int minute;
-        int hour;
+        int8_t second;
+        int8_t minute;
+        int8_t hour;
         unsigned long previous_millis;
         unsigned long current_millis;
         bool is_running;
@@ -14,8 +14,8 @@ class Chronometer {
     public:
         Chronometer();
         ~Chronometer();
-        void curUp(int* pos_cursor);
-        void curDown(int* pos_cursor);
+        void curUp(int8_t* pos_cursor);
+        void curDown(int8_t* pos_cursor);
         void toggle();
         void secondCount();
         void print(LiquidCrystal* lcd) const;
